add details page
