+++ app/components/candidates/View.vue
@@ -1,0 +1,23 @@
+<template>
+    <div>
+        <div class="table-container">
+            <md-card>
+                <md-card-header>
+                    <div class="md-title">Candidate</div>
+                    <div class="md-subhead">{{ $route.params.address }}</div>
+                </md-card-header>
+
+                <!--md-card-content>
+                    
+                </md-card-content-->
+
+                <!--md-card-actions>
+                    <md-button>Action</md-button>
+                    <md-button>Action</md-button>
+                </md-card-actions-->
+            </md-card>
+        </div>
+    </div>
+</template>
+<script>
+</script>
