+++ package.json
@@ -42,6 +42,7 @@
         "vue": "^2.5.16",
         "vue-loader": "^14.2.2",
         "vue-material": "^1.0.0-beta-8",
+        "vue-router": "^3.0.1",
         "vue-template-compiler": "^2.5.16",
         "webpack": "^4.5.0",
         "webpack-cli": "^2.0.14"
